+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### BUG FIXES
 
 * Deployment with a topology parsing error remains in initial status ([GH-283](https://github.com/ystia/yorc/issues/283)
+* Interface name is not retrieved from custom command Rest request ([GH-287](https://github.com/ystia/yorc/issues/287)
 
 ## 3.2.0-M1 (January 28, 2019)